requirements.txt
# requirements.txt
yt-dlp==2025.1.12
faster-whisper==1.0.3
webvtt-py==0.5.1
tqdm==4.66.5
unidecode==1.3.8
